Create file for CopyToClipboardButton
