+++ src/App.tsx
@@ -1,48 +1,7 @@
 import React, { useState, useEffect } from "react";
 import NewItemForm from "./components/NewItemForm";
 import ItemsList from "./components/ItemsList";
-
-interface ItemsListProps {
-  items: Array<Item>;
-  removeItem: (index: number) => void;
-  updateItem: (index: number, item: Item) => void;
-}
-
-interface CopyToClipboardProps {
-  items: Array<Item>;
-}
-
-const CopyToClipboardButton: React.FC<CopyToClipboardProps> = ({ items }) => {
-  const [text, setText] = useState("Copy to clipboard");
-
-  function copyToClipboard() {
-    let textArea = document.createElement("textarea");
-    textArea.value = itemsToString();
-    document.body.appendChild(textArea);
-    textArea.select();
-    document.execCommand("copy");
-    document.body.removeChild(textArea);
-
-    setText("Copied!");
-    setTimeout(() => setText("Copy to clipboard"), 2000);
-  }
-
-  function itemsToString() {
-    return items.map((item) => `${item.quantity} x ${item.name}`).join("\n");
-  }
-
-  return (
-    <div className="" style={{ width: "175px" }}>
-      <button
-        type="button"
-        className="btn btn-secondary btn-block"
-        onClick={copyToClipboard}
-      >
-        {text}
-      </button>
-    </div>
-  );
-};
+import CopyToClipboardButton from "./components/CopyToClipboardButton"
 
 export interface Item {
   name: string;
